2nd attempt at re-connecting on disconnect.
When the connection to the Blue-iris server fails, we'll most
likely detect this in the cmd() function which traps the error
and returns false.

In short poll, we call cmd() but then never checked the return
so we didn't know the connection had disconnected.  Now, check
the return value and if false, attempt to re-connect.

Still not sure we're handling all of this correctly.